Move each block type into separate file :truck:
