+++ Src/Notion.Client/Models/Blocks/BulletedListItemBlock.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class BulletedListItemBlock : BlockBase
+    {
+        public override BlockType Type => BlockType.BulletedListItem;
+
+        [JsonProperty("bulleted_list_item")]
+        public Info BulletedListItem { get; set; }
+
+        public class Info
+        {
+            public IEnumerable<RichTextBase> Text { get; set; }
+            public IEnumerable<BlockBase> Children { get; set; }
+        }
+    }
+}
