+++ Src/Notion.Client/Models/Blocks/HeadingThreeeBlock.cs
@@ -1,0 +1,20 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class HeadingThreeeBlock : BlockBase
+    {
+        public override BlockType Type => BlockType.Heading_3;
+
+        [JsonProperty("heading_3")]
+        public Info Heading_3 { get; set; }
+
+        public override bool HasChildren => false;
+
+        public class Info
+        {
+            public IEnumerable<RichTextBase> Text { get; set; }
+        }
+    }
+}
