+++ Src/Notion.Client/Models/Blocks/UnsupportedBlock.cs
@@ -1,0 +1,7 @@
+﻿namespace Notion.Client
+{
+    public class UnsupportedBlock : BlockBase
+    {
+        public override BlockType Type => BlockType.Unsupported;
+    }
+}
